requirements.txt
kivy==2.1.0
requests>=2.28.0
pyjnius>=1.4.2
plyer>=2.1.0
gTTS>=2.4.0
pygame>=2.5.2
python-dotenv>=1.0.0
Pillow>=9.5.0
chardet>=5.2.0
urllib3>=1.26.0
cryptography>=42.0.0  # <- Añade esta línea